Add validation of runner inputs
We've forgotten an input for `subdir-darwin-arm64`. This is fixed in
another PR, but it raises the question of validation -- given that this
set of inputs is required by the design of the action, it is essential
complexity resulting in a necessary foot-gun. This commit implements a
validation to (hopefully) avoid such mistakes going forward.

Ideally, we might list all possible values of the platform and arch
types and ensure we have inputs for all of them in one go as a unit
test. We don't do this for two reasons:

1. Listing the values of a union type is not possible
2. We don't want 100% coverage of these values anyway, and the values we
   do want aren't even a simple cross-product of constants

Together, this makes generation and/or hard-coding of desired values a
both non-starters.

Since all we actually care about are the possible values we'll see on
GitHub runners, we can get what we need by running the validation _on_
GitHub runners and using the values found by calling `node` right there.

At that point, it's only a question of which runners we care to run this
particular validation on. As a matter of pragmatism, I'm going with the
integration test matrix set, which is each `-latest` of the major OSes.